replace `toSearch` with `toQueryClause`
the terminology was bad. it is not a search, but instead, a component of
a search. search related information like sorting and paging are not
included as apart of our input. it's at a higher level. this library
only acts on the query dsl portion of the actual search itself. this is
very misleading and would be sure to confuse consumers.